+++ src/eu/ttbox/geoping/domain/geotrack/GeoTrackHelper.java
@@ -60,8 +60,8 @@ public class GeoTrackHelper {
         geoTrack.setAccuracy(accuracyIdx > -1 ? cursor.getInt(accuracyIdx) : -1);
         geoTrack.setAltitude(altitudeIdx > -1 ? cursor.getDouble(altitudeIdx) : -1);
 
-        geoTrack.setBearing(bearingIdx > -1 ? cursor.getFloat(bearingIdx) : -1);
-        geoTrack.setSpeed(speedIdx > -1 ? cursor.getFloat(speedIdx) : -1);
+        geoTrack.setBearing(bearingIdx > -1 ? cursor.getInt(bearingIdx) : -1);
+        geoTrack.setSpeed(speedIdx > -1 ? cursor.getInt(speedIdx) : -1);
 
         return geoTrack;
     }
